Send all galaxy initialization log directly to stderr and/or stdout (#197)
This commit removes the config_db.log from our start-up procedure.
Previously, this file captured only stdout from run_galaxy_config.sh
and in case of failure was copied to the galaxy pvc so that it could be
read after the pod restart.  With this change everything is visible from the pod
logs, UNLESS it's redirected by the startup command in the helm chart and, of course, if you can get it before k8s destroys them.